refactor: utilsWeb 폴더를 제거하고 유틸 함수의 경로를 utils 폴더로 통일
