+++ src/view/web/layers/footer/footer.js
@@ -1,5 +1,5 @@
 import footerContents from "./footer.html?raw";
-import { appendContents } from "../../utilsWeb/elementCreator.js";
+import { appendContents } from "../../../../utils/elementCreator.js";
 import "./footer.css";
 
 const createFooter = () => {
