+++ src/view/web/layers/winningLottoBox/createNumbersInput.js
@@ -1,5 +1,5 @@
 import { LOTTO_RULE } from "../../../../constants/lotto.js";
-import { setNumberInput } from "../../utilsWeb/input.js";
+import { setNumberInput } from "../../../../utils/templates.js";
 
 const createNumbersInput = () => {
   return Array.from({ length: LOTTO_RULE.LOTTO_LENGTH })
